[hasher] fix integration test after git mv
